requirements.txt
torch
torchvision
numpy
pandas
matplotlib
seaborn
scikit-learn
tqdm
opencv-python
Pillow
jupyter
